test(ai): split treasury_planner_test forecasting group into own file (#2924)
treasury_planner_test.dart exceeded the 1000 non-comment line repo-lint
ceiling (1009). Move the partial-fill-aware forecasting and speculative-bid
groups into treasury_planner_forecasting_test.dart; both files now sit well
under the limit. No behavior change; all 22 tests still pass.
